:recycle: Split some of the stuff out of main.ts.
This makes main.ts a little smaller and easier to maintain.
